Extract fixture service/app.rb
